requirements.txt
openai==1.30.1
anthropic==0.28.0
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn==0.30.1
pandas==2.2.2
pydantic==2.7.1
